whispers-python-lambdas/requirements.txt
psycopg2-binary
boto3
amazon-dax-client
Faker
